LRAC-4544 move
