+++ modules/apps/analytics/analytics-settings-impl/src/main/java/com/liferay/analytics/settings/internal/security/auth/verifier/AnalyticsSecurityAuthVerifier.java
@@ -12,10 +12,10 @@
  * details.
  */
 
-package com.liferay.analytics.security.internal.security.auth.verifier;
+package com.liferay.analytics.settings.internal.security.auth.verifier;
 
-import com.liferay.analytics.security.internal.constants.AnalyticsSecurityConstants;
 import com.liferay.analytics.settings.configuration.AnalyticsConfiguration;
+import com.liferay.analytics.settings.security.constants.AnalyticsSecurityConstants;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONUtil;
 import com.liferay.portal.kernel.log.Log;
